requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.89
botocore==1.34.89
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
frozenlist==1.4.1
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
imageio==2.34.0
ipykernel==4.8.2
ipython==8.23.0
ipython-genutils==0.2.0
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
wsproto==1.2.0
yarl==1.9.4
flask
waitress
flask_cors
datrie
opencv-python